Consolidate blockwise FP8 MoE grouped MM
