+++ torchao/prototype/moe_training/fp8_grouped_mm.py
@@ -4,10 +4,39 @@
 # This source code is licensed under the BSD 3-Clause license found in the
 # LICENSE file in the root directory of this source tree.
 
+from dataclasses import dataclass
+from enum import Enum
 from typing import Optional
 
 import torch
 
+from torchao.float8.config import e4m3_dtype
+from torchao.prototype.blockwise_fp8_training.deepgemm_grouped_kernels import (
+    can_use_deepgemm_grouped_training,
+    deepgemm_blockwise_scaled_grouped_mm,
+    deepgemm_blockwise_scaled_grouped_mm_wgrad,
+    prepare_deepgemm_wgrad_plan,
+)
+from torchao.prototype.blockwise_fp8_training.deepgemm_metadata import (
+    DeepGemmGroupedOffsetPlan,
+    build_deepgemm_grouped_offset_plan,
+)
+from torchao.prototype.blockwise_fp8_training.grouped_kernels import (
+    emulated_blockwise_scaled_grouped_mm,
+)
+from torchao.prototype.blockwise_fp8_training.grouped_weight_quant import (
+    triton_fp8_blockwise_weight_quant_grouped_dgrad_rhs,
+    triton_fp8_blockwise_weight_quant_grouped_forward_rhs,
+)
+from torchao.prototype.blockwise_fp8_training.kernels import (
+    BLOCKWISE_1X128_SCALING_TYPE,
+    BLOCKWISE_128X128_SCALING_TYPE,
+    _is_row_major,
+    _scaling_type_value,
+    triton_fp8_blockwise_act_quant_lhs,
+    triton_fp8_blockwise_act_quant_rhs,
+    triton_fp8_blockwise_act_quant_transposed_lhs,
+)
 from torchao.prototype.moe_training.kernels import (
     triton_fp8_colwise_3d_scale_and_cast,
     triton_fp8_per_group_colwise_scales_dual,
@@ -16,9 +45,11 @@ from torchao.prototype.moe_training.kernels import (
 )
 from torchao.prototype.moe_training.utils import (
     _is_column_major,
+    conditional_nostrict_trace,
     pad_token_groups,
     unpad_token_groups,
 )
+from torchao.quantization.quantize_.common import KernelPreference
 
 
 def _to_fp8_rowwise_then_scaled_grouped_mm(
@@ -317,3 +348,557 @@ class _Float8GroupedMM(torch.autograd.Function):
             use_fast_accum=True,
         )
         return grad_A, grad_B.transpose(-2, -1), None, None, None, None
+
+
+class _GroupedMMBackendKind(str, Enum):
+    """Grouped GEMM backend selected for the FP8 MoE training op."""
+
+    DEEPGEMM = "deepgemm"
+    EMULATED = "emulated"
+
+
+class _GroupedMMBackend:
+    """Backend-specific quantization and grouped GEMM implementation."""
+
+    kind: _GroupedMMBackendKind
+
+    def quantize_forward_rhs(
+        self,
+        B_t: torch.Tensor,
+        block_size: int,
+        dtype: torch.dtype,
+    ) -> tuple[torch.Tensor, torch.Tensor]:
+        raise NotImplementedError
+
+    def quantize_dgrad_rhs(
+        self,
+        B_t: torch.Tensor,
+        block_size: int,
+        dtype: torch.dtype,
+    ) -> tuple[torch.Tensor, torch.Tensor]:
+        raise NotImplementedError
+
+    def grouped_mm(
+        self,
+        a: torch.Tensor,
+        b: torch.Tensor,
+        a_s: torch.Tensor,
+        scale_recipe_a: int,
+        b_s: torch.Tensor,
+        scale_recipe_b: int,
+        offs: torch.Tensor,
+        out_dtype: torch.dtype,
+        block_size: int,
+    ) -> torch.Tensor:
+        raise NotImplementedError
+
+    def wgrad(
+        self,
+        padded_grad_output: torch.Tensor,
+        padded_a: torch.Tensor,
+        group_end_offsets: torch.Tensor,
+        out_dtype: torch.dtype,
+        block_size: int,
+        dtype: torch.dtype,
+    ) -> torch.Tensor:
+        raise NotImplementedError
+
+
+class _EmulatedGroupedMMBackend(_GroupedMMBackend):
+    """TorchAO emulated backend using PyTorch grouped-mm operand layouts."""
+
+    kind = _GroupedMMBackendKind.EMULATED
+
+    def quantize_forward_rhs(
+        self,
+        B_t: torch.Tensor,
+        block_size: int,
+        dtype: torch.dtype,
+    ) -> tuple[torch.Tensor, torch.Tensor]:
+        # The shared direct quantizer writes forward RHS as row-major
+        # (E, N, K) data with (E, N_blocks, K_blocks) scales. Transpose views
+        # convert that to the RHS contract required by torch._grouped_mm and
+        # torch._scaled_grouped_mm for output = A @ B_t: (E, K, N) data with
+        # (E, K_blocks, N_blocks) scales.
+        q, scale = triton_fp8_blockwise_weight_quant_grouped_forward_rhs(
+            B_t,
+            block_size=block_size,
+            dtype=dtype,
+        )
+        return q.transpose(-2, -1), scale.transpose(-2, -1)
+
+    def quantize_dgrad_rhs(
+        self,
+        B_t: torch.Tensor,
+        block_size: int,
+        dtype: torch.dtype,
+    ) -> tuple[torch.Tensor, torch.Tensor]:
+        # The shared direct quantizer writes dgrad RHS as row-major
+        # (E, K, N) data with (E, K_blocks, N_blocks) scales. Transpose views
+        # convert that to the RHS contract required by torch._grouped_mm and
+        # torch._scaled_grouped_mm for grad_output @ weight: (E, N, K) data with
+        # (E, N_blocks, K_blocks) scales.
+        q, scale = triton_fp8_blockwise_weight_quant_grouped_dgrad_rhs(
+            B_t,
+            block_size=block_size,
+            dtype=dtype,
+        )
+        return q.transpose(-2, -1), scale.transpose(-2, -1)
+
+    def grouped_mm(
+        self,
+        a: torch.Tensor,
+        b: torch.Tensor,
+        a_s: torch.Tensor,
+        scale_recipe_a: int,
+        b_s: torch.Tensor,
+        scale_recipe_b: int,
+        offs: torch.Tensor,
+        out_dtype: torch.dtype,
+        block_size: int,
+    ) -> torch.Tensor:
+        return emulated_blockwise_scaled_grouped_mm(
+            a,
+            b,
+            a_s,
+            scale_recipe_a,
+            b_s,
+            scale_recipe_b,
+            offs,
+            out_dtype,
+            block_size,
+        )
+
+    def wgrad(
+        self,
+        padded_grad_output: torch.Tensor,
+        padded_a: torch.Tensor,
+        group_end_offsets: torch.Tensor,
+        out_dtype: torch.dtype,
+        block_size: int,
+        dtype: torch.dtype,
+    ) -> torch.Tensor:
+        # For expert e, wgrad is [N, M_e] @ [M_e, K] -> [N, K]. PyTorch's
+        # 2D x 2D grouped_mm contract represents all experts as two 2D tensors:
+        #   grad_output_t: [N, M], row-major
+        #   A:             [M, K], column-major
+        # `group_end_offsets` partitions their shared M dimension into M_e.
+        grad_output_t_fp8, grad_output_t_scale = (
+            triton_fp8_blockwise_act_quant_transposed_lhs(
+                padded_grad_output.contiguous(),
+                block_size=block_size,
+                dtype=dtype,
+            )
+        )
+        A_rhs_fp8, A_rhs_scale = triton_fp8_blockwise_act_quant_rhs(
+            padded_a.contiguous(),
+            block_size=block_size,
+            dtype=dtype,
+        )
+        return emulated_blockwise_scaled_grouped_mm(
+            grad_output_t_fp8,
+            A_rhs_fp8,
+            grad_output_t_scale,
+            _scaling_type_value(BLOCKWISE_1X128_SCALING_TYPE),
+            A_rhs_scale,
+            _scaling_type_value(BLOCKWISE_1X128_SCALING_TYPE),
+            group_end_offsets,
+            out_dtype,
+            block_size,
+        )
+
+
+@dataclass(frozen=True)
+class _DeepGemmGroupedMMBackend(_GroupedMMBackend):
+    """DeepGEMM backend plus the offset metadata shared by its kernels."""
+
+    kind = _GroupedMMBackendKind.DEEPGEMM
+    offset_plan: DeepGemmGroupedOffsetPlan
+
+    def quantize_forward_rhs(
+        self,
+        B_t: torch.Tensor,
+        block_size: int,
+        dtype: torch.dtype,
+    ) -> tuple[torch.Tensor, torch.Tensor]:
+        # DeepGEMM forward consumes RHS as (E, N, K), with K contiguous and
+        # scales as (E, N_blocks, K_blocks). This quantizer writes that
+        # layout directly, avoiding a dispatch-time transpose/copy.
+        return triton_fp8_blockwise_weight_quant_grouped_forward_rhs(
+            B_t,
+            block_size=block_size,
+            dtype=dtype,
+        )
+
+    def quantize_dgrad_rhs(
+        self,
+        B_t: torch.Tensor,
+        block_size: int,
+        dtype: torch.dtype,
+    ) -> tuple[torch.Tensor, torch.Tensor]:
+        # DeepGEMM dgrad consumes RHS as (E, K, N), with N contiguous and
+        # scales as (E, K_blocks, N_blocks). This quantizer writes that
+        # layout directly, avoiding a dispatch-time transpose/copy.
+        return triton_fp8_blockwise_weight_quant_grouped_dgrad_rhs(
+            B_t,
+            block_size=block_size,
+            dtype=dtype,
+        )
+
+    def grouped_mm(
+        self,
+        a: torch.Tensor,
+        b: torch.Tensor,
+        a_s: torch.Tensor,
+        scale_recipe_a: int,
+        b_s: torch.Tensor,
+        scale_recipe_b: int,
+        offs: torch.Tensor,
+        out_dtype: torch.dtype,
+        block_size: int,
+    ) -> torch.Tensor:
+        return deepgemm_blockwise_scaled_grouped_mm(
+            a,
+            b,
+            a_s,
+            scale_recipe_a,
+            b_s,
+            scale_recipe_b,
+            offs,
+            out_dtype,
+            block_size,
+            offset_plan=self.offset_plan,
+        )
+
+    def wgrad(
+        self,
+        padded_grad_output: torch.Tensor,
+        padded_a: torch.Tensor,
+        group_end_offsets: torch.Tensor,
+        out_dtype: torch.dtype,
+        block_size: int,
+        dtype: torch.dtype,
+    ) -> torch.Tensor:
+        # DeepGEMM computes the same [N, M_e] @ [M_e, K] wgrad. Its API calls
+        # this K-grouped because the expert-dependent M_e is the GEMM
+        # reduction/K extent. `k_grouped_fp8_gemm_nt_contiguous` takes two flat
+        # expert-major buffers. The quantizer
+        # concatenates row-major [N, M_e] blocks in expert order for the LHS
+        # and row-major [K, M_e] blocks in expert order for the RHS. The flat
+        # segment lengths are therefore [N * M_0, ..., N * M_{E-1}] and
+        # [K * M_0, ..., K * M_{E-1}], respectively.
+        wgrad_plan = prepare_deepgemm_wgrad_plan(
+            padded_grad_output,
+            padded_a,
+            self.offset_plan,
+            block_size,
+            dtype,
+        )
+        assert wgrad_plan is not None, (
+            "DeepGEMM backend requires block-aligned group sizes for wgrad"
+        )
+        return deepgemm_blockwise_scaled_grouped_mm_wgrad(
+            wgrad_plan.lhs,
+            wgrad_plan.rhs,
+            self.offset_plan,
+            out_dtype,
+            block_size,
+        )
+
+
+_EMULATED_GROUPED_MM_BACKEND = _EmulatedGroupedMMBackend()
+
+
+def _select_fp8_blockwise_grouped_mm_backend(
+    kernel_preference: KernelPreference,
+    A: torch.Tensor,
+    out_dtype: torch.dtype,
+    block_size: int,
+    group_end_offsets: torch.Tensor,
+    *,
+    original_group_end_offsets: Optional[torch.Tensor] = None,
+    padded_group_start_offsets: Optional[torch.Tensor] = None,
+    num_rows: Optional[int] = None,
+) -> _GroupedMMBackend:
+    """Select the grouped GEMM backend for one forward/backward pass.
+
+    ``KernelPreference.EMULATED`` always selects the TorchAO emulated backend.
+    ``KernelPreference.AUTO`` and ``KernelPreference.DEEPGEMM`` select DeepGEMM
+    when the optional dependency exposes both M-grouped and K-grouped training
+    kernels, the input is on CUDA SM90+, ``out_dtype`` is bf16, ``block_size``
+    is 128, and every expert group is block-aligned. Unsupported cases fail;
+    emulation must be selected explicitly.
+    When DeepGEMM is selected, the returned backend owns the offset/layout plan
+    reused by forward, dgrad, and wgrad. The autograd function saves this
+    backend, so wgrad cannot independently choose a different layout or kernel.
+    """
+
+    if kernel_preference == KernelPreference.EMULATED:
+        return _EMULATED_GROUPED_MM_BACKEND
+
+    assert kernel_preference in (
+        KernelPreference.AUTO,
+        KernelPreference.DEEPGEMM,
+    ), "kernel_preference must be AUTO, DEEPGEMM, or EMULATED"
+    if not can_use_deepgemm_grouped_training(A, out_dtype, block_size):
+        raise RuntimeError(
+            f"KernelPreference.{kernel_preference.name} could not select a supported "
+            "non-emulated FP8 blockwise grouped GEMM backend. DeepGEMM requires "
+            "CUDA SM90+, bfloat16 output, block_size=128, and an installed "
+            "`deep_gemm` package with M-grouped and K-grouped training kernels. "
+            "Select KernelPreference.EMULATED to use emulation."
+        )
+
+    groups_block_aligned_by_construction = original_group_end_offsets is not None
+    offset_plan = build_deepgemm_grouped_offset_plan(
+        group_end_offsets,
+        original_group_end_offsets=original_group_end_offsets,
+        padded_group_start_offsets=padded_group_start_offsets,
+        num_rows=num_rows,
+        groups_block_aligned_by_construction=groups_block_aligned_by_construction,
+    )
+    if not offset_plan.groups_are_block_aligned(block_size):
+        raise RuntimeError(
+            "DeepGEMM FP8 blockwise grouped GEMM requires every expert token "
+            f"count to be divisible by {block_size}. Enable "
+            "pad_token_groups_for_grouped_mm or select KernelPreference.EMULATED."
+        )
+
+    return _DeepGemmGroupedMMBackend(offset_plan=offset_plan)
+
+
+@conditional_nostrict_trace
+def _to_fp8_blockwise_then_scaled_grouped_mm(
+    A: torch.Tensor,
+    B_t: torch.Tensor,
+    offs: Optional[torch.Tensor] = None,
+    out_dtype: Optional[torch.dtype] = torch.bfloat16,
+    float8_dtype: torch.dtype = e4m3_dtype,
+    block_size: int = 128,
+    pad_token_groups_for_grouped_mm: bool = True,
+    kernel_preference: KernelPreference = KernelPreference.AUTO,
+) -> torch.Tensor:
+    """
+    Differentiable FP8 blockwise grouped matrix multiplication.
+
+    A has shape (M, K). B_t has shape (E, K, N), transposed and in
+    per-expert column-major layout.
+
+    ``kernel_preference`` selects one backend for the complete autograd
+    operation. ``DEEPGEMM`` explicitly selects DeepGEMM, ``AUTO`` selects
+    DeepGEMM as the only available non-emulated backend, and ``EMULATED`` uses
+    PyTorch grouped-mm layouts and kernels. ``AUTO`` and ``DEEPGEMM`` fail when
+    DeepGEMM is unsupported instead of falling back to emulation. The selected
+    backend is reused for forward, dgrad, and wgrad.
+    """
+    assert block_size == 128, "Only block_size=128 is supported"
+    assert kernel_preference in (
+        KernelPreference.AUTO,
+        KernelPreference.DEEPGEMM,
+        KernelPreference.EMULATED,
+    ), "kernel_preference must be AUTO, DEEPGEMM, or EMULATED"
+    return _Float8BlockwiseGroupedMM.apply(
+        A,
+        B_t,
+        offs,
+        out_dtype,
+        float8_dtype,
+        block_size,
+        pad_token_groups_for_grouped_mm,
+        kernel_preference,
+    )
+
+
+@conditional_nostrict_trace
+def _to_fp8_blockwise_then_emulated_scaled_grouped_mm(
+    A: torch.Tensor,
+    B_t: torch.Tensor,
+    offs: Optional[torch.Tensor] = None,
+    out_dtype: Optional[torch.dtype] = torch.bfloat16,
+    float8_dtype: torch.dtype = e4m3_dtype,
+    block_size: int = 128,
+    pad_token_groups_for_grouped_mm: bool = True,
+) -> torch.Tensor:
+    """
+    Differentiable FP8 blockwise grouped matrix multiplication using an emulated GEMM.
+    """
+    return _to_fp8_blockwise_then_scaled_grouped_mm(
+        A,
+        B_t,
+        offs,
+        out_dtype,
+        float8_dtype,
+        block_size,
+        pad_token_groups_for_grouped_mm,
+        KernelPreference.EMULATED,
+    )
+
+
+class _Float8BlockwiseGroupedMM(torch.autograd.Function):
+    @staticmethod
+    def forward(
+        ctx,
+        A: torch.Tensor,
+        B_t: torch.Tensor,
+        offs: Optional[torch.Tensor] = None,
+        out_dtype: Optional[torch.dtype] = torch.bfloat16,
+        float8_dtype: torch.dtype = e4m3_dtype,
+        block_size: int = 128,
+        pad_token_groups_for_grouped_mm: bool = True,
+        kernel_preference: KernelPreference = KernelPreference.AUTO,
+    ) -> torch.Tensor:
+        assert A.ndim == 2, "A must be 2D"
+        assert B_t.ndim == 3, "B_t must be 3D"
+        assert offs is not None and offs.dtype == torch.int32, "offs must be int32"
+        assert A.dtype in (torch.bfloat16, torch.float32), "A must be bf16 or fp32"
+        assert B_t.dtype in (torch.bfloat16, torch.float32), "B_t must be bf16 or fp32"
+        assert A.shape[-1] == B_t.shape[-2], (
+            f"shape {A.shape} and {B_t.shape} are not compatible"
+        )
+        assert A.shape[-1] % block_size == 0 and B_t.shape[-1] % block_size == 0, (
+            f"K and N must be divisible by block_size={block_size}"
+        )
+        assert _is_row_major(A), "A must be row-major"
+        assert _is_column_major(B_t), "B_t must be per-expert column-major"
+
+        num_tokens = A.shape[0]
+        padded_group_start_offsets = None
+        padded_group_end_offsets = offs
+        if pad_token_groups_for_grouped_mm:
+            padded_A, padded_group_start_offsets, padded_group_end_offsets = (
+                pad_token_groups(A, offs, alignment_size=block_size)
+            )
+        else:
+            padded_A = A
+
+        backend = _select_fp8_blockwise_grouped_mm_backend(
+            kernel_preference,
+            A,
+            out_dtype,
+            block_size,
+            padded_group_end_offsets,
+            original_group_end_offsets=offs
+            if pad_token_groups_for_grouped_mm
+            else None,
+            padded_group_start_offsets=padded_group_start_offsets,
+            num_rows=padded_A.shape[0],
+        )
+
+        A_fp8, A_scale = triton_fp8_blockwise_act_quant_lhs(
+            padded_A.contiguous(),
+            block_size=block_size,
+            dtype=float8_dtype,
+        )
+        B_t_fp8, B_t_scale = backend.quantize_forward_rhs(
+            B_t,
+            block_size,
+            float8_dtype,
+        )
+        out = backend.grouped_mm(
+            A_fp8,
+            B_t_fp8,
+            A_scale,
+            _scaling_type_value(BLOCKWISE_1X128_SCALING_TYPE),
+            B_t_scale,
+            _scaling_type_value(BLOCKWISE_128X128_SCALING_TYPE),
+            padded_group_end_offsets,
+            out_dtype,
+            block_size,
+        )
+        if pad_token_groups_for_grouped_mm:
+            out = unpad_token_groups(
+                out,
+                offs,
+                padded_group_start_offsets,
+                num_tokens,
+                alignment_size=block_size,
+            )
+
+        ctx.save_for_backward(
+            padded_A,
+            B_t,
+            offs,
+            padded_group_start_offsets,
+            padded_group_end_offsets,
+        )
+        ctx.out_dtype = out_dtype
+        ctx.float8_dtype = float8_dtype
+        ctx.block_size = block_size
+        ctx.pad_token_groups_for_grouped_mm = pad_token_groups_for_grouped_mm
+        ctx.num_tokens = num_tokens
+        ctx.backend = backend
+        return out
+
+    @staticmethod
+    def backward(ctx, grad_output: torch.Tensor):
+        (
+            padded_A,
+            B_t,
+            original_group_end_offsets,
+            padded_group_start_offsets,
+            padded_group_end_offsets,
+        ) = ctx.saved_tensors
+        out_dtype = ctx.out_dtype
+        float8_dtype = ctx.float8_dtype
+        block_size = ctx.block_size
+        pad_token_groups_for_grouped_mm = ctx.pad_token_groups_for_grouped_mm
+        num_tokens = ctx.num_tokens
+        backend = ctx.backend
+
+        if pad_token_groups_for_grouped_mm:
+            padded_grad_output, _, _ = pad_token_groups(
+                grad_output,
+                original_group_end_offsets,
+                alignment_size=block_size,
+            )
+        else:
+            padded_grad_output = grad_output
+
+        grad_output_fp8, grad_output_scale = triton_fp8_blockwise_act_quant_lhs(
+            padded_grad_output.contiguous(),
+            block_size=block_size,
+            dtype=float8_dtype,
+        )
+        B_fp8, B_scale = backend.quantize_dgrad_rhs(
+            B_t,
+            block_size,
+            float8_dtype,
+        )
+        grad_A = backend.grouped_mm(
+            grad_output_fp8,
+            B_fp8,
+            grad_output_scale,
+            _scaling_type_value(BLOCKWISE_1X128_SCALING_TYPE),
+            B_scale,
+            _scaling_type_value(BLOCKWISE_128X128_SCALING_TYPE),
+            padded_group_end_offsets,
+            out_dtype,
+            block_size,
+        )
+        if pad_token_groups_for_grouped_mm:
+            grad_A = unpad_token_groups(
+                grad_A,
+                original_group_end_offsets,
+                padded_group_start_offsets,
+                num_tokens,
+                alignment_size=block_size,
+            )
+
+        grad_B = backend.wgrad(
+            padded_grad_output,
+            padded_A,
+            padded_group_end_offsets,
+            out_dtype,
+            block_size,
+            float8_dtype,
+        )
+        return (
+            grad_A,
+            grad_B.transpose(-2, -1),
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+        )
